TC#60: Refactored changes
