+++ package.json
@@ -17,9 +17,8 @@
     "axios": "^1.7.8",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
+    "date-fns": "^3.6.0",
     "embla-carousel-autoplay": "^8.5.1",
-    "embla-carousel-react": "^8.5.1",
-    "date-fns": "^4.1.0",
     "embla-carousel-react": "^8.5.1",
     "lucide-react": "^0.461.0",
     "next": "14.2.16",
