--- src/actions/blog.ts
@@ -1,8 +1,0 @@
-import { axiosApi } from '@/shared/lib/axios-api';
-import { Blogs } from '@/shared/types/blog.types';
-
-export const getBlogs = async (page: string | null) => {
-  const { data: blogs } = await axiosApi.get<Blogs>(`/news${page && `?page=${page}`}`);
-
-  return blogs;
-};
