--- src/app/blog/page.module.css
@@ -1,7 +1,0 @@
-.blogs {
-  @apply mb-14;
-}
-
-.pagination {
-  @apply mb-20;
-}
