--- src/shared/components/shared/blogs/blogs.module.css
@@ -1,3 +1,0 @@
-.blogs {
-  @apply grid sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-x-5 gap-y-[40px] justify-center;
-}
