--- src/shared/lib/generate-pages.ts
@@ -1,25 +1,0 @@
-export const generatePages = (currentPage: number, totalPages: number) => {
-  const pages: (number | string)[] = [];
-
-  if (totalPages <= 5) {
-    for (let i = 1; i <= totalPages; i++) {
-      pages.push(i);
-    }
-  } else {
-    if (currentPage > 3) {
-      pages.push(1, '...');
-    }
-    const start = Math.max(2, currentPage - 1);
-    const end = Math.min(totalPages - 1, currentPage + 1);
-
-    for (let i = start; i <= end; i++) {
-      pages.push(i);
-    }
-
-    if (currentPage < totalPages - 2) {
-      pages.push('...', totalPages);
-    }
-  }
-
-  return pages;
-};
