+++ src/shared/lib/index.ts
@@ -1,3 +1,3 @@
-export { generatePages } from './generate-pages';
 export { axiosApi } from './axios-api';
 export { cn } from './utils';
+export { usePagination } from './hooks/usePagination';
